+++ R/countrycode.R
@@ -104,7 +104,7 @@ countrycode <- function(sourcevar, origin, destination, warn = TRUE, nomatch = N
             "cctld", "country.name", "country.name.de", "country.name.fr", "country.name.it",
             "cowc", "cown", "dhs", "ecb", "eurostat", "fao", "fips", "gaul",
             "genc2c", "genc3c", "genc3n", "gwc", "gwn", "imf", "ioc", "iso2c",
-            "iso3c", "iso3n", "p4c", "p4n", "un", "un_m49", "unicode.symbol",
+            "iso3c", "iso3n", "p5c", "p5n", "p4c", "p4n", "un", "un_m49", "unicode.symbol",
             "unhcr", "unpd", "vdem", "wb", "wb_api2c", "wb_api3c", "wvs",
             "country.name.en.regex", "country.name.de.regex", "country.name.fr.regex", "country.name.it.regex")
         attr(dictionary, "origin_regex") <- c("country.name.de.regex",