+++ tests/testthat/test-basic.R
@@ -6,6 +6,12 @@ test_that('valid iso3c to country.name works', {
     expect_equal(name_of(c('USA', 'CAN')), c('United States', 'Canada'))
 })
 
+
+test_that("Issue #252: new code polity5", {
+    expect_equal(countrycode("United States", "country.name", "p5c"), "USA")
+    expect_equal(countrycode("United States", "country.name", "p5n"), 2)
+})
+
 test_that('invalid iso3c to country.name returns NA', {
     name_of <- function(iso3c_code) countrycode(iso3c_code, 'iso3c', 'country.name', warn = FALSE)
     expect_equal(name_of('BAD'), NA_character_)