+++ NEWS.md
@@ -1,5 +1,6 @@
 # countrycode 1.4.0.9000
 
+* New codes: Polity V "p5c", "p5n"
 * Many regex improvements
 * Several minor bug fixes